+++ src/lib/types/entry.php
@@ -1,6 +1,31 @@
 <?php
 
 /**
+ * Get entry type css class, it will split the entry type id
+ * on slash and take the last part of the id.
+ *
+ * @param  int $id
+ * @param  string $type
+ *
+ * @return string
+ */
+function papi_get_entry_type_css_class( $id = 0, $type = 'post' ) {
+	$entry_type = papi_get_entry_type_id( $id, $type );
+
+	if ( empty( $entry_type ) ) {
+		return '';
+	}
+
+	$parts = explode( '/', $entry_type );
+
+	if ( empty( $parts ) || empty( $parts[0] ) ) {
+		return '';
+	}
+
+	return array_pop( $parts );
+}
+
+/**
  * Count entry types in the database for the
  * given entry type.
  *