+++ tests/cases/lib/core/template-test.php
@@ -27,7 +27,7 @@ class Papi_Lib_Core_Template_Test extends WP_UnitTestCase {
 		$this->assertEmpty( papi_body_class( [] ) );
 
 		update_post_meta( $this->post_id, papi_get_page_type_key(), 'simple-page-type' );
-		$this->assertSame( [ 'simple-page-type' ], papi_body_class( [] ) );
+		$this->assertSame( ['simple-page-type'], papi_body_class( [] ) );
 	}
 
 	public function test_papi_get_template_file_name() {