+++ CHANGELOG.md
@@ -17,6 +17,7 @@
 * Added `display` option to box options in order to control if the box should be displayed or not.
 * Added support for repeaters inside repeaters.
 * Added support for render html in `Publish box` with `publish_box` method.
+* Added `papi_get_entry_type_css_class` to get the css class that is added to body for a entry type.
 
 ### Changed